+++ NEWS.md
@@ -1,5 +1,10 @@
 # gittins (development version)
 
+* `bmab_gi_multiple()` (breaking changes):
+  * Now accepts alpha/beta or Sigma/n state inputs.
+  * Therefore `bmab_gi_multiple_ab()` is no longer needed.
+  * Output is now a data frame with a row for each starting state. The matrix output is attached as 
+  attributes.
 * Streamline code and docs.
 * Add GPL3 licence.